backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
firebase-admin
pyjwt
langgraph
langchain-core
langchain-openai